+++ src/app/cart/components/cart-item/cart-item.component.ts
@@ -1,20 +1,17 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 
+import { ProductModel } from '../../../products/models/product.model';
+
 @Component({
   selector: 'app-cart-item',
   templateUrl: './cart-item.component.html',
   styleUrls: ['./cart-item.component.scss']
 })
-export class CartItemComponent implements OnInit {
-  @Input() cartItem;
+export class CartItemComponent {
+  @Input() cartItem: ProductModel;
   @Output() deleteItem = new EventEmitter();
 
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-  onDeleteItem() {
+  onDeleteItem(): void {
     this.deleteItem.emit(this.cartItem);
   }